[release-v0.33] 33차 버스 출발

+++ src/domains/job/recruitment/detail/page/RecruitmentDetailPage.jsx
@@ -58,6 +58,9 @@ const RecruitmentDetailPage = () => {
   }
 
   const handleDeleteClick = async () => {
+    const confirmed = window.confirm('삭제 후 복원이 불가능합니다. 정말 삭제하시겠습니까?')
+    if (!confirmed) return
+
     try {
       await deleteRecruitmentPost(id)
       ToastService.success('성공적으로 삭제되었습니다.')
@@ -203,7 +206,7 @@ const RecruitmentDetailPage = () => {
         <MobileShare label={data.viewCount} icon={viewPink} textColor='text-main-pink' />
       </div>
       <div
-        className='block mt-4 mb-10 text-sm text-left break-words whitespace-pre-line sm:text-base'
+        className='block mt-4 mb-10 text-[16px] leading-relaxed text-left break-words whitespace-pre-line'
         dangerouslySetInnerHTML={{ __html: DOMPurify.sanitize(data.text) }}
       />
       <LastFormLine />
